Refactored code and remove old functions
