+++ cmd/manager/main.go
@@ -118,11 +118,6 @@ func main() {
 		os.Exit(1)
 	}
 
-	if err := specialresource.SetupCallbacks(); err != nil {
-		log.Error(err, "")
-		os.Exit(1)
-	}
-
 	// Setup all Controllers
 	if err := controller.AddToManager(mgr); err != nil {
 		log.Error(err, "")
