--- pkg/controller/specialresource/tools.go
@@ -1,1 +1,0 @@
-package specialresource
