+++ recipes/nvidia-gpu/config/Makefile
@@ -1,6 +1,6 @@
 
 nvidia-gpu: deploy 
-	-kubectl delete configmap -n $(NAMESPACE) -l specialresource.openshift.io/config=true
+	-kubectl delete configmap -n $(NAMESPACE) 0000-$@
 	kubectl create configmap 0000-$@ -n $(NAMESPACE) --from-file=recipes/$@/manifests
-	kubectl patch configmap 0000-$@ -n $(NAMESPACE)  --patch "$(cat recipes/$@/config/patch.yaml)" 
+	kubectl patch configmap 0000-$@ -n $(NAMESPACE)  --patch '$(shell cat recipes/$@/config/patch.json)'
